+++ src/main/java/com/erp/dto/SalesChartDTO.java
@@ -12,5 +12,5 @@ import java.util.List;
 @ToString
 public class SalesChartDTO {
     private final List<String> labels;   // X축: 날짜(일/주/월/년)
-    private final List<Integer> values;  // Y축: 매출 합계
+    private final List<Long> values;  // Y축: 매출 합계
 }
